+++ frontend/src/components/EditDestination.jsx
@@ -1,0 +1,40 @@
+import {
+  FocusLock,
+  IconButton,
+  Popover,
+  PopoverArrow,
+  PopoverCloseButton,
+  PopoverContent,
+  PopoverTrigger, useDisclosure
+} from "@chakra-ui/react";
+import {EditIcon} from "@chakra-ui/icons";
+import PopoverForm from "./destinationComponents/PopoverForm.jsx";
+import {useRef} from "react";
+
+export default function EditDestination() {
+  const { onOpen, onClose, isOpen } = useDisclosure()
+  const firstFieldRef = useRef(null)
+  return (
+    <div style={{textAlign: "center"}}>
+      <Popover
+        isOpen={isOpen}
+        initialFocusRef={firstFieldRef}
+        onOpen={onOpen}
+        onClose={onClose}
+        placement='right'
+        closeOnBlur={false}
+      >
+        <PopoverTrigger>
+          <IconButton size='sm' icon={<EditIcon/>}/>
+        </PopoverTrigger>
+        <PopoverContent p={5}>
+          <FocusLock returnFocus persistentFocus={false}>
+            <PopoverArrow/>
+            <PopoverCloseButton/>
+            <PopoverForm firstFieldRef={firstFieldRef} onCancel={onClose}/>
+          </FocusLock>
+        </PopoverContent>
+      </Popover>
+    </div>
+  )
+}
